requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
discord-ext-voice_recv @ git+https://github.com/imayhaveborkedit/discord-ext-voice-recv@02a0782db6a1e154edbcfe252004f8fdd4bf539f
discord.py==2.5.2
edge-tts==7.0.2
frozenlist==1.6.2
idna==3.10
multidict==6.4.4
numpy==2.2.6
propcache==0.3.1
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.3
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
